+++ BDSKApplication.m
@@ -86,8 +86,8 @@
         [[NSNotificationCenter defaultCenter] postNotificationName:BDSKFlagsChangedNotification object:self];
 }
 
-- (unsigned int)currentModifierFlags {
-    unsigned int flags = 0;
+- (NSUInteger)currentModifierFlags {
+    NSUInteger flags = 0;
     UInt32 currentKeyModifiers = GetCurrentKeyModifiers();
     if (currentKeyModifiers & cmdKey)
         flags |= NSCommandKeyMask;
@@ -126,15 +126,15 @@
     NSMenu *windowsMenu = [self windowsMenu];
     NSWindowController *windowController = [aWindow windowController];
     NSWindowController *mainWindowController = [[[[aWindow windowController] document] windowControllers] objectAtIndex:0];
-    int numberOfItems = [windowsMenu numberOfItems];
-    int itemIndex = [windowsMenu indexOfItemWithTarget:aWindow andAction:@selector(makeKeyAndOrderFront:)];
+    NSInteger numberOfItems = [windowsMenu numberOfItems];
+    NSInteger itemIndex = [windowsMenu indexOfItemWithTarget:aWindow andAction:@selector(makeKeyAndOrderFront:)];
     
     if (itemIndex != -1) {
         NSMenuItem *item = [windowsMenu itemAtIndex:itemIndex];
         NSString *title = [item title];
         
         if ([windowController document] == nil) {
-            int idx = numberOfItems;
+            NSInteger idx = numberOfItems;
             while (idx--) {
                 NSMenuItem *anItem = [windowsMenu itemAtIndex:idx];
                 if ([anItem isSeparatorItem] ||
@@ -156,8 +156,8 @@
         } else if ([windowController isEqual:mainWindowController]) {
             NSMutableArray *subitems = [NSMutableArray array];
             NSMenuItem *anItem;
-            int idx = numberOfItems;
-            int nextIndex = numberOfItems;
+            NSInteger idx = numberOfItems;
+            NSInteger nextIndex = numberOfItems;
             
             while (idx--) {
                 anItem = [windowsMenu itemAtIndex:idx];
@@ -208,8 +208,8 @@
                 [windowsMenu insertItem:[NSMenuItem separatorItem] atIndex:nextIndex];
             
         } else {
-            int mainIndex = [windowsMenu indexOfItemWithTarget:[mainWindowController window] andAction:@selector(makeKeyAndOrderFront:)];
-            int idx = mainIndex;
+            NSInteger mainIndex = [windowsMenu indexOfItemWithTarget:[mainWindowController window] andAction:@selector(makeKeyAndOrderFront:)];
+            NSInteger idx = mainIndex;
             
             [item setIndentationLevel:1];
             
@@ -233,7 +233,7 @@
 }
 
 - (void)addWindowsItem:(NSWindow *)aWindow title:(NSString *)aString filename:(BOOL)isFilename {
-    int itemIndex = [[self windowsMenu] indexOfItemWithTarget:aWindow andAction:@selector(makeKeyAndOrderFront:)];
+    NSInteger itemIndex = [[self windowsMenu] indexOfItemWithTarget:aWindow andAction:@selector(makeKeyAndOrderFront:)];
     
     [super addWindowsItem:aWindow title:aString filename:isFilename];
     
@@ -250,7 +250,7 @@
 - (void)removeWindowsItem:(NSWindow *)aWindow {
     [super removeWindowsItem:aWindow];
     
-    int idx = [[self windowsMenu] numberOfItems];
+    NSInteger idx = [[self windowsMenu] numberOfItems];
     BOOL wasSeparator = YES;
     
     while (idx--) {
@@ -269,7 +269,7 @@
 
 - (NSArray *)orderedDocuments {
     NSMutableArray *orderedDocuments = [[[super orderedDocuments] mutableCopy] autorelease];
-    int i = [orderedDocuments count];
+    NSInteger i = [orderedDocuments count];
     
     while (i--)
         if ([[orderedDocuments objectAtIndex:i] isKindOfClass:[BibDocument class]] == NO)
